weatherchatbot/src/weatherchatbot/requirements.txt
crewai
python-dotenv
ipykernel
agentops 
crewai[tools]
fastapi
uvicorn
streamlit